pyproject.toml
[project]
name = "linkedin-extractor"
version = "0.1.0"
description = ""
authors = [
    {name = "Kunal7069",email = "jaink7069@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "linkedin_extractor", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
